requirements.txt
gradio
requests
Pillow
catboost
numpy
joblib
scikit-learn